Add a way to view a historic deployment
Note: not linked to yet, ideally it would be linked from the recent
deployments list on an application page, as this is where you're
most likely to want to see exactly what a previous release involved.
But that would make the way we link 'releases' a bit inconsistent
and I want to give that some thought and will add in a follow up
commit.

It's useful to be able to see what a historic deployment involved,
especially compared to the release previous to this one. The common
use case here is identifying where a deployment may have had an
effect on performance/errirs/some other system metric, previously
this could only by done via the console.

This exposes the version, time, env, previous version, allowing us
to link out to a diff of exactly what was requested.