Link to historic deployment views, rather than GH
May be contentious. Mostly when we references a release_XXX tag
we're talking about a previous deployment, now we have pages for
them in the release app we should link to them.

These deployment pages still have a link to the tag on GH, but
I'm assuming this change may annoy some people, it would be
interesting to hear common uses cases.

Certainly there's a distinction between a tag/treeish that we
often refer to as a 'release' (including an application/repo)
and a deployment, which is comprised of a time, a 'release',
an application and an environment.